Add new quicklaunch-report command to compare launchwizard & streams
To find out if launchwizard quickstart entries are up-to-date, it's
useful to compare the the AMIs found in the launchwizard are equal to
the AMIs found in streams.
This new report command does exactly that and outputs a table for
comparison.
It can be used with:

python -m aws_marketplace_ubuntu_scraper quicklaunch-report