update marketplace scraping due to many changes to the rendering of the seller marketplace listings
The new UI for marketplace listings isn't always the one served but in my testing is used for
90% of requests. As such I have changed to from using BS to parse HTML to using Selenium to
scrape an authenticated API request for all marketplace listings.

This commit also adds parsing of the Ubuntu based 2p listings owned by AWS. These are usually the
Ubunto Pro listings but also the the Ubuntu AMIs maintained by AWS.

This allows us to see the serial/age of the AMIs used in the marketplace listings

See https://pastebin.ubuntu.com/p/jPXrKdzfvr/ for output as of 20211222.