Prepare quicklaunch-report command for usage in automation
quicklaunch-report will be used in automation. For that, it's useful
to
a) show only the entries where an update is needed
b) detect if updates are needed at all

Both cases are solved with this commit.
The new --needs-update-only flag can be used to only list entries that
need an update.
The command will now have 2 as returncode if updates are needed.